Move touchables to universal-components
