+++ packages/components/index.js
@@ -2,15 +2,11 @@
 
 import BlackToAlpha from './src/placeCard/assets/black-to-alpha-vertical.png';
 
-export {
-  default as TouchableWithoutFeedback,
-} from './src/TouchableWithoutFeedback';
 export { default as TripInput } from './src/tripInput/TripInput';
 export { default as Select } from './src/select/Select';
 export { default as Modal } from './src/modal/Modal';
 export { default as TripTypeButton } from './src/tripTypeButton/TripTypeButton';
 export { default as TripTypeSwitch } from './src/tripTypeSwitch/TripTypeSwitch';
-export { default as ExtendedTouchable } from './src/ExtendedTouchable';
 export {
   default as PassengersButton,
 } from './src/passengersButton/PassengersButton';
