+++ packages/components/src/passengersButton/__tests__/PassengersButton.test.js
@@ -2,9 +2,9 @@
 
 import * as React from 'react';
 import { shallow, render, fireEvent } from 'react-native-testing-library';
+import { TouchableWithoutFeedback } from '@kiwicom/universal-components';
 
 import PassengersButton from '../PassengersButton';
-import TouchableWithoutFeedback from '../../TouchableWithoutFeedback';
 
 const adults = 2;
 const infants = 2;
