+++ packages/universal-components/src/ExtendedTouchable/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export { default as ExtendedTouchable } from './ExtendedTouchable';
